Reorganize typed didcomm service models
