+++ did_core/did_doc/src/schema/service/typed/aip1.rs
