+++ did_core/did_doc/src/schema/service/typed/didcommv2.rs
@@ -1,9 +1,12 @@
-use serde::Serialize;
+use display_as_json::Display;
+use serde::{Deserialize, Serialize};
+use typed_builder::TypedBuilder;
 use url::Url;
 
 use crate::schema::{
     service::{
-        extra_fields::{didcommv2::ExtraFieldsDidCommV2, ServiceAcceptType, ServiceKeyKind},
+        service_accept_type::ServiceAcceptType,
+        service_key_kind::ServiceKeyKind,
         typed::{ServiceType, TypedService},
     },
     types::uri::Uri,
@@ -51,5 +54,25 @@ impl ServiceDidCommV2 {
 
     pub fn extra(&self) -> &ExtraFieldsDidCommV2 {
         self.service.extra()
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default, Display, TypedBuilder)]
+#[serde(rename_all = "camelCase")]
+#[serde(deny_unknown_fields)]
+pub struct ExtraFieldsDidCommV2 {
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    routing_keys: Vec<ServiceKeyKind>,
+    #[serde(default)]
+    accept: Vec<ServiceAcceptType>,
+}
+
+impl ExtraFieldsDidCommV2 {
+    pub fn routing_keys(&self) -> &[ServiceKeyKind] {
+        self.routing_keys.as_ref()
+    }
+
+    pub fn accept(&self) -> &[ServiceAcceptType] {
+        self.accept.as_ref()
     }
 }
