+++ did_core/did_doc/src/schema/service/typed/legacy.rs
@@ -2,7 +2,7 @@ use display_as_json::Display;
 use serde::{Deserialize, Serialize};
 use typed_builder::TypedBuilder;
 
-use crate::schema::service::extra_fields::ServiceKeyKind;
+use crate::schema::service::service_key_kind::ServiceKeyKind;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Default, Display, TypedBuilder)]
 #[serde(rename_all = "camelCase")]
