+++ did_core/did_doc/src/schema/service/typed/mod.rs
@@ -1,3 +1,4 @@
+pub mod aip1;
 pub mod didcommv1;
 pub mod didcommv2;
