+++ did_core/did_methods/did_peer/src/peer_did/numalgos/traits.rs
@@ -1,5 +1,4 @@
 use did_doc::schema::did_doc::DidDocument;
-use did_doc_sov::extra_fields::ExtraFieldsSov;
 use did_parser::Did;
 
 use crate::{
